Bug修正：1、界面单位；2、临界值计算；3、钻杆立根长度；4、起下钻次数；6、充值后未刷新

+++ SteelWire/SteelWire.Business/Database/WorkConfig.cs
@@ -31,7 +31,7 @@ namespace SteelWire.Business.Database
         public string DrillingDifficulty { get; set; }
         public decimal TripShallowHeight { get; set; }
         public decimal TripDeepHeight { get; set; }
-        public int TripCount { get; set; }
+        public decimal TripCount { get; set; }
         public decimal BushingWeight { get; set; }
         public decimal BushingLength { get; set; }
         public decimal BushingHeight { get; set; }
